+++ models/marketInfo.go
@@ -53,7 +53,7 @@ type MarketInfo struct {
 	TakerFee                float64
 	Slippage                float64
 	PricePrecision          float64
-	QuantityPrecision       float64
+	QuantityPrecision       int
 	MinimumOrderSize        float64
 	MaxLeverage             float64
 	BulkCancelSupported     bool
